move daffodil-lib schema package to lib subpackage
