+++ daffodil-core/src/main/scala/org/apache/daffodil/dsom/AnnotatedSchemaComponent.scala
@@ -22,11 +22,13 @@ import org.apache.daffodil.exceptions.Assert
 import org.apache.daffodil.xml.XMLUtils
 import org.apache.daffodil.xml.NS
 import org.apache.daffodil.equality._
-import org.apache.daffodil.schema.annotation.props.PropertyLookupResult
-import org.apache.daffodil.schema.annotation.props.NotFound
-import org.apache.daffodil.schema.annotation.props.Found
-import org.apache.daffodil.schema.annotation.props.FindPropertyMixin
 import org.apache.daffodil.api.WarnID
+import org.apache.daffodil.lib.schema.annotation.props
+import org.apache.daffodil.lib.schema.annotation.props.FindPropertyMixin
+import org.apache.daffodil.lib.schema.annotation.props.Found
+import org.apache.daffodil.lib.schema.annotation.props.NotFound
+import org.apache.daffodil.lib.schema.annotation.props.PropertyLookupResult
+
 import scala.collection.mutable
 
 /**
@@ -84,7 +86,7 @@ trait ResolvesLocalProperties
     val fa = formatAnnotation
     val opt = fa.justThisOneProperties.get(pname)
     val optFound = opt.map { case (value, location) => Found(value, location, pname, false) }
-    val res = optFound.getOrElse { NotFound(Seq(fa.annotatedSC), Nil, pname) }
+    val res = optFound.getOrElse { props.NotFound(Seq(fa.annotatedSC), Nil, pname) }
     res
   }
 }
