+++ daffodil-core/src/main/scala/org/apache/daffodil/dsom/CompiledExpression.scala
@@ -19,13 +19,15 @@ package org.apache.daffodil.dsom
 
 import org.apache.daffodil.dpath._
 import org.apache.daffodil.dpath.NodeInfo.PrimType
+
 import scala.xml.NamespaceBinding
 import org.apache.daffodil.xml.NamedQName
-import java.lang.{ Long => JLong, Boolean => JBoolean }
-import org.apache.daffodil.schema.annotation.props.Found
+
+import java.lang.{Long => JLong, Boolean => JBoolean}
 import org.apache.daffodil.util.DPathUtil
 import org.apache.daffodil.infoset.DataValue.DataValuePrimitive
 import org.apache.daffodil.BasicComponent
+import org.apache.daffodil.lib.schema.annotation.props.Found
 
 object ExpressionCompilers extends ExpressionCompilerClass {
   override val String = new ExpressionCompiler[String]
