+++ daffodil-core/src/main/scala/org/apache/daffodil/dsom/DFDLFormatAnnotation.scala
@@ -19,10 +19,9 @@ package org.apache.daffodil.dsom
 
 import scala.collection.immutable.ListMap
 import scala.xml.Node
-
 import org.apache.daffodil.api.WarnID
 import org.apache.daffodil.exceptions.Assert
-import org.apache.daffodil.schema.annotation.props.LookupLocation
+import org.apache.daffodil.lib.schema.annotation.props.LookupLocation
 import org.apache.daffodil.xml.NS
 import org.apache.daffodil.xml.NoNamespace
 import org.apache.daffodil.xml.RefQName
