+++ daffodil-core/src/main/scala/org/apache/daffodil/dsom/EscapeSchemeRefMixin.scala
@@ -18,8 +18,8 @@
 package org.apache.daffodil.dsom
 
 import org.apache.daffodil.api.WarnID
-import org.apache.daffodil.schema.annotation.props.Found
-import org.apache.daffodil.schema.annotation.props.NotFound
+import org.apache.daffodil.lib.schema.annotation.props.Found
+import org.apache.daffodil.lib.schema.annotation.props.NotFound
 
 trait EscapeSchemeRefMixin { self: Term =>
   /**
