+++ daffodil-core/src/main/scala/org/apache/daffodil/grammar/primitives/SequenceCombinator.scala
@@ -20,9 +20,9 @@ package org.apache.daffodil.grammar.primitives
 import org.apache.daffodil.dsom._
 import org.apache.daffodil.grammar.Gram
 import org.apache.daffodil.grammar.Terminal
+import org.apache.daffodil.lib.schema.annotation.props.SeparatorSuppressionPolicy
 import org.apache.daffodil.processors.parsers._
 import org.apache.daffodil.processors.unparsers._
-import org.apache.daffodil.schema.annotation.props.SeparatorSuppressionPolicy
 import org.apache.daffodil.util.Misc
 import org.apache.daffodil.util.Maybe
 import org.apache.daffodil.util.MaybeInt
