+++ daffodil-runtime1/src/main/scala/org/apache/daffodil/processors/parsers/SeparatedSequenceChildParseResultHelper.scala
@@ -17,7 +17,7 @@
 package org.apache.daffodil.processors.parsers
 
 import org.apache.daffodil.exceptions.Assert
-import org.apache.daffodil.schema.annotation.props.EmptyElementParsePolicy
+import org.apache.daffodil.lib.schema.annotation.props.EmptyElementParsePolicy
 import org.apache.daffodil.processors.ElementRuntimeData
 import org.apache.daffodil.processors.ModelGroupRuntimeData
