+++ daffodil-runtime1/src/main/scala/org/apache/daffodil/processors/parsers/UnseparatedSequenceChildParseResultHelper.scala
@@ -16,8 +16,8 @@
  */
 package org.apache.daffodil.processors.parsers
 
+import org.apache.daffodil.lib.schema.annotation.props.EmptyElementParsePolicy
 import org.apache.daffodil.processors.ElementRuntimeData
-import org.apache.daffodil.schema.annotation.props.EmptyElementParsePolicy
 import org.apache.daffodil.processors.ModelGroupRuntimeData
 
 /**
